fix: split tests into separate files
